execs as modules moved to tests script

+++ tests/run-tests.sh
@@ -4,7 +4,23 @@ sitepack=$(python -c 'import sys; \
            sitedirs=[p for p in sys.path if p.endswith("site-packages")]; \
            print sitedirs[0]')
 
-[[ ! -L $sitepack/argo_egi_connectors ]] && \
+if [[ ! -L $sitepack/argo_egi_connectors ]]
+then
     ln -s $PWD/modules $sitepack/argo_egi_connectors
+fi
+
+(cd bin && \
+for f in [a-z]*.py
+do
+    link=$(echo $f | sed "s/-/_/g") 
+    if [[ ! -L $link ]]
+    then
+        ln -s $f $link > /dev/null
+    fi
+done && \
+if [[ ! -e __init__.py ]]
+then 
+    touch __init__.py
+fi)
 
 coverage run --source=tests -m unittest2 discover tests && coverage xml
